+++ src/main/java/io/github/qudtlib/maven/rdfio/ExcludeFilter.java
@@ -1,7 +1,0 @@
-package io.github.qudtlib.maven.rdfio;
-
-public class ExcludeFilter extends AbstractFilter {
-    public ExcludeFilter() {
-        super(false);
-    }
-}