+++ src/main/java/io/github/qudtlib/maven/rdfio/Filter.java
@@ -1,7 +1,0 @@
-package io.github.qudtlib.maven.rdfio;
-
-import org.apache.jena.rdf.model.Model;
-
-public interface Filter {
-    void filter(Model model);
-}