+++ src/main/java/io/github/qudtlib/maven/rdfio/IncludeFilter.java
@@ -1,7 +1,0 @@
-package io.github.qudtlib.maven.rdfio;
-
-public class IncludeFilter extends AbstractFilter {
-    public IncludeFilter() {
-        super(true);
-    }
-}