+++ src/main/java/io/github/qudtlib/maven/rdfio/filter/ExcludeFilter.java
@@ -1,0 +1,7 @@
+package io.github.qudtlib.maven.rdfio.filter;
+
+public class ExcludeFilter extends AbstractPredicateFilter {
+    public ExcludeFilter() {
+        super(false);
+    }
+}