+++ src/main/java/io/github/qudtlib/maven/rdfio/filter/IncludeFilter.java
@@ -1,0 +1,7 @@
+package io.github.qudtlib.maven.rdfio.filter;
+
+public class IncludeFilter extends AbstractPredicateFilter {
+    public IncludeFilter() {
+        super(true);
+    }
+}