+++ src/main/java/io/github/qudtlib/maven/rdfio/product/Product.java
@@ -1,0 +1,10 @@
+package io.github.qudtlib.maven.rdfio.product;
+
+import org.apache.jena.rdf.model.Model;
+import org.apache.maven.plugin.MojoExecutionException;
+
+public interface Product {
+    void process(Model model) throws MojoExecutionException;
+
+    String describe();
+}