+++ src/main/java/io/github/qudtlib/maven/rdfio/product/SingleFile.java
@@ -1,9 +1,13 @@
-package io.github.qudtlib.maven.rdfio;
+package io.github.qudtlib.maven.rdfio.product;
 
+import io.github.qudtlib.maven.rdfio.filter.Filters;
+import io.github.qudtlib.maven.rdfio.filter.IncludeExcludePatterns;
 import org.apache.jena.rdf.model.Model;
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugin.logging.Log;
 import org.apache.maven.plugins.annotations.Parameter;
 
-public class SingleFile {
+public class SingleFile implements Product {
     @Parameter private IncludeExcludePatterns input;
 
     @Parameter private String outputFile;
@@ -12,6 +16,8 @@ public class SingleFile {
     private boolean skip;
 
     @Parameter private Filters filters;
+
+    private Log log;
 
     public IncludeExcludePatterns getInput() {
         return input;
@@ -34,9 +40,19 @@ public class SingleFile {
         return "Check{" + "shapes='" + input + '\'' + '}';
     }
 
-    public void filter(Model inputGraph) {
+    public void process(Model inputGraph) throws MojoExecutionException {
         if (this.filters != null) {
+            this.filters.setLog(log);
             this.filters.filter(inputGraph);
         }
+    }
+
+    public void setLog(Log log) {
+        this.log = log;
+    }
+
+    @Override
+    public String describe() {
+        return getOutputFile();
     }
 }