+++ CHANGELOG.md
@@ -7,19 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- EachFile product type, allowing for a set of input files to be processed individually and optionally changed in-place.
+- SparqlUpdate filter, allowing for deleting and adding triples based on a sparql update query.
+
 ## [1.1.2] - 2024-11-08
 
 ## [1.1.1] - 2024-11-08
 ### Fixed
+
 - Fixed bug which caused multiline strings to be written with a '\r' appended to each line.
 
 ## [1.1.0] - 2024-11-05
 ### Added
+
 - Allow filtering of RDF content
 
 ## 1.0.0 - 2024-11-03
 ### Added
-- Initial version of the plugin offering just making a single RDF file as the combination of multiple files
+
+- Initial version of the plugin offering just making a single RDF file as the combination of multiple files.
 
 [Unreleased]: https://github.com/qudtlib/rdfio-maven-plugin/compare/v1.1.2...HEAD
 [1.1.2]: https://github.com/qudtlib/rdfio-maven-plugin/compare/v1.1.1...v1.1.2